fix error with reallocations when parsing string expressions
String expressions were parsed and grabbed a pointer to the underlying
buffer directly, prior to a code path which **coud** trigger reallocations.
As a result, sometime reallocation would cause the previous pointer to
point to invalid memory, resulting in garbled reads of data. This fix
switches to using handles, which are then later converted into the pointer.